requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
azure-functions==1.23.0
azure-functions-durable==1.3.2
certifi==2025.7.9
charset-normalizer==3.4.2
frozenlist==1.7.0
furl==2.1.4
idna==3.10
importlib_metadata==8.7.0
MarkupSafe==3.0.2
multidict==6.6.3
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orderedmultidict==1.0.1
pillow==11.3.0
propcache==0.3.2
pyodbc==5.2.0
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
